# openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/agent/test/ingest": {
      "post": {
        "operationId": "AgentController_testIngest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestIngestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "[TEST] Ingest an episode into the knowledge graph",
        "tags": [
          "Agent"
        ]
      }
    },
    "/agent/chat": {
      "post": {
        "operationId": "AgentController_chat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Chat with the agent (pre-fetch + knowledge search)",
        "tags": [
          "Agent"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hoard API",
    "description": "Epic memory augmentation app!",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TestIngestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "content": {
            "type": "string",
            "minLength": 1
          },
          "resolveCoreferences": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "content",
          "resolveCoreferences"
        ]
      },
      "ChatDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}